+++ lib/checkother.cpp
@@ -1612,7 +1612,7 @@ void CheckOther::checkConstPointer()
         }
         if (std::find(nonConstPointers.begin(), nonConstPointers.end(), p) == nonConstPointers.end()) {
             const Token *start = (p->isArgument()) ? p->scope()->bodyStart : p->nameToken()->next();
-            const int indirect = p->isArray() ? p->dimensions().size() : 1;
+            const int indirect = p->isArray() ? static_cast<int>(p->dimensions().size()) : 1;
             if (isVariableChanged(start, p->scope()->bodyEnd, indirect, p->declarationId(), false, mSettings, mTokenizer->isCPP()))
                 continue;
             if (p->isArgument() && p->typeStartToken() && p->typeStartToken()->isSimplifiedTypedef() && !(Token::simpleMatch(p->typeEndToken(), "*") && !p->typeEndToken()->isSimplifiedTypedef()))
