+++ lib/programmemory.cpp
@@ -848,7 +848,7 @@ static std::unordered_map<std::string, BuiltinLibraryFunction> createBuiltinLibr
         double value = args[0].isFloatValue() ? args[0].floatValue : args[0].intvalue;
         ValueFlow::Value v;
         combineValueProperties(args[0], args[1], &v);
-        v.floatValue = std::ldexp(value, args[1].isFloatValue() ? args[1].floatValue : args[1].intvalue);
+        v.floatValue = std::ldexp(value, args[1].isFloatValue() ? args[1].floatValue : static_cast<int>(args[1].intvalue));
         v.valueType = ValueFlow::Value::ValueType::FLOAT;
         return v;
     };
